+++ R/data.R
@@ -1,0 +1,15 @@
+#' Lego color data.
+#'
+#' A data frame containing Lego color data.
+#'
+#' @format A data frame with 12 columns and 167 rows.
+"legocolors"
+
+#' Lego color palettes.
+#'
+#' A named list of nine Lego color palettes based on \code{legocolors$material}.
+#' Because these palettes represent a meaningful segmentation of Lego material types,
+#' they are not intended to represent useful color palettes for mapping data values to colors.
+#'
+#' @format A named list of nine color palettes.
+"legopals"